+++ app/src/main/java/com/example/pazzly/domain/entity/KorakPoKorak.java
@@ -5,7 +5,7 @@ public class KorakPoKorak extends Game{
     public KorakPoKorak() {
     }
 
-    public KorakPoKorak(int rounds, int maxPointsPerRound, int minimalPointsPerRound, int durationPerRound) {
+    public KorakPoKorak(int rounds, int maxPointsPerRound, int minimalPointsPerRound, double durationPerRound) {
         super(rounds, maxPointsPerRound, minimalPointsPerRound, durationPerRound, 1);
     }
 }
